+++ apps/frontend-control/next.config.mjs
@@ -42,6 +42,10 @@ let nextConfig = {
   },
 }
 
+nextConfig.transpilePackages = Array.from(
+  new Set([...(nextConfig.transpilePackages ?? []), 'formik'])
+)
+
 if (process.env.NODE_ENV !== 'test') {
   const withPWA = withPWAInit(
     getNextPWAConfig({ NODE_ENV: process.env.NODE_ENV })